+++ iivs/dhm/data/intensity/base.py
@@ -1,6 +1,12 @@
 from __future__ import annotations
 
-__all__ = ("IntensityFloatSequence", "IntensityImageSequence", "IntensitySequence")
+__all__ = (
+    "IntensityFileFolder",
+    "IntensityFileList",
+    "IntensityFloatSequence",
+    "IntensityImageSequence",
+    "IntensitySequence",
+)
 
 from pathlib import Path
 from typing import TYPE_CHECKING, ClassVar